change variable reading type from int to long

+++ src/main/java/fr/insee/genesis/domain/service/contextualvariable/previous/ContextualPreviousVariableJsonService.java
@@ -160,7 +160,7 @@ public class ContextualPreviousVariableJsonService implements ContextualPrevious
                 return jsonParser.getText();
             }
             case VALUE_NUMBER_INT -> {
-                return jsonParser.getIntValue();
+                return jsonParser.getLongValue();
             }
             case VALUE_NUMBER_FLOAT -> {
                 return jsonParser.getDoubleValue();
